Fix territory id type: uint not ushort
Current Dalamud exposes TerritoryType as uint and TerritoryChanged as
Action<uint>. Caught by CI on first Windows build.

+++ HistoryEntry.cs
@@ -26,5 +26,5 @@ public readonly record struct HistoryEntry(
     byte Stain,              // current dye id. Redyed: the NEW dye.
     byte FromStain,          // Redyed only: the previous dye id.
     ulong HouseId,           // which house this happened in (for multi-house disambiguation)
-    ushort TerritoryId,
+    uint TerritoryId,
     bool WhileAway);         // true if detected on entry — i.e. changed since your last visit
